+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Initial release of nf-core/predictorthologs, created with the [nf-core](http://n
 - Added option of using sourmash to search instead of DIAMOND. This stays in hash-land by searching hashes (provided by `--hashes` or found by `--differential_hash_expression`) directly into a sourmash sequence bloom tree index database
 - Added option for bam deduplication, if you wish to skip deduplication step add the `-skip_remove_duplicates_bam` flag
 - Added ability to search DIAMOND for hashes that were unassigned from sourmash ([#79](https://github.com/czbiohub/nf-predictorthologs/pull/79))
+- Add version printing for sencha, and sourmash, update versions in environment.yml
 
 ### `Fixed`